requirements.txt
langchain
deeplake
openai
tiktoken
streamlit
streamlit_chat
python-dotenv
bs4
requests
pypdf
unstructured
tabulate
pdf2image